+++ CHANGELOG.md
@@ -16,3 +16,18 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - 
+
+## [0.2.0] - 2026-06-16
+
+> **Important**: This is a 0.x release. The API and functions are **unstable** and may change in future versions.
+
+### Added
+- 
+
+### Changed
+- Hashtable is now prebuilt
+- Hashtable is now 16MB instead of 32MB
+- set_faster function deleted
+
+### Fixed
+-